extract NodeActionWarning component
