+++ src/app/base/components/node/NodeActionWarning/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NodeActionWarning";
